Fixed a bunch of NPE.
