+++ src/main/java/com/github/evillootlye/caves/ticks/Dynamics.java
@@ -28,14 +28,4 @@ public class Dynamics {
         tickables.get(tick.getTickLevel()).add(tick);
     }
 
-    public enum TickLevel {
-        WORLD(6300), ENTITY(4);
-
-        public final int ticks;
-
-        TickLevel(int ticks) {
-            this.ticks = ticks;
-        }
-    }
-
 }
