+++ src/main/java/com/github/evillootlye/caves/ticks/TickLevel.java
@@ -1,0 +1,12 @@
+package com.github.evillootlye.caves.ticks;
+
+// TODO: Configurable?
+public enum TickLevel {
+    WORLD(6300), ENTITY(4);
+
+    public final int ticks;
+
+    TickLevel(int ticks) {
+        this.ticks = ticks;
+    }
+}
