+++ src/main/java/com/github/evillootlye/caves/ticks/Tickable.java
@@ -2,5 +2,5 @@ package com.github.evillootlye.caves.ticks;
 
 public interface Tickable {
     void tick();
-    Dynamics.TickLevel getTickLevel();
+    TickLevel getTickLevel();
 }
